Crate snowbridge_pallet_system_v2

Governance API for controlling the Ethereum side of the bridge

Extrinsics

Governance

Call::upgrade: Upgrade the Gateway contract on Ethereum.

Call::set_operating_mode: Set the operating mode of the Gateway contract

Polkadot-native tokens on Ethereum

Tokens deposited on AssetHub pallet can be bridged to Ethereum as wrapped ERC20 tokens. As a prerequisite, the token should be registered first.

Call::register_token: Register a token location as a wrapped ERC20 contract on Ethereum.

Re-exports

pub use weights::*;

pub use pallet::*;

Modules

api

Helpers for implementing runtime api

pallet

The pallet module in each FRAME pallet hosts the most important items needed to construct this pallet.

weights

Autogenerated weights for snowbridge_system

Constants

LOG_TARGET

Traits

BenchmarkHelper

Type Aliases

AccountIdOf